sched/fair: Cast normalized_sysctl_sched_base_slice to u64
Clang doesn't like it.
Fixes warning:
../kernel/sched/fair.c:934:13: warning: comparison of distinct pointer types ('typeof (se->slice) *' (aka 'unsigned long long *') and 'typeof (normalized_sysctl_sched_base_slice) *' (aka 'unsigned int *')) [-Wcompare-distinct-pointer-types]
934 |                 quantum = min(se->slice, normalized_sysctl_sched_base_slice);
    |                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:862:19: note: expanded from macro 'min'
862 | #define min(x, y)       __careful_cmp(x, y, <)
    |                         ^~~~~~~~~~~~~~~~~~~~~~
2025-06-14T10:20:18.4922075Z ../include/linux/kernel.h:853:24: note: expanded from macro '__careful_cmp'
853 |         __builtin_choose_expr(__safe_cmp(x, y), \
    |                               ^~~~~~~~~~~~~~~~
../include/linux/kernel.h:843:4: note: expanded from macro '__safe_cmp'
843 |                 (__typecheck(x, y) && __no_side_effects(x, y))
    |                  ^~~~~~~~~~~~~~~~~
../include/linux/kernel.h:829:29: note: expanded from macro '__typecheck'
829 |                 (!!(sizeof((typeof(x) *)1 == (typeof(y) *)1)))
    |                            ~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~
1 warning generated.

+++ kernel/sched/fair.c
@@ -931,7 +931,7 @@ static inline void set_protect_slice(struct sched_entity *se)
 	if (sched_feat(RUN_TO_PARITY))
 		quantum = cfs_rq_min_slice(cfs_rq_of(se));
 	else
-		quantum = min(se->slice, normalized_sysctl_sched_base_slice);
+		quantum = min(se->slice,(u64)normalized_sysctl_sched_base_slice);
 
 	if (quantum != se->slice)
 		se->vlag = min(se->deadline, se->vruntime + calc_delta_fair(quantum, se));
